factura a medias

+++ Entidad/Cliente.cs
@@ -24,7 +24,7 @@ namespace Entidad
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string CorreoElectronico { get; set; }
-        public decimal Telefono { get; set; }
+        public int Telefono { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Factura> Factura { get; set; }
